+++ pkg/cmd/hibernate/clusterclaim/cmd.go
@@ -38,7 +38,7 @@ func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOSt
 		},
 		Run: func(cmd *cobra.Command, args []string) {
 			cmdutil.CheckErr(o.complete(cmd, args))
-			cmdutil.CheckErr(o.validate())
+			cmdutil.CheckErr(o.validate(cmd))
 			cmdutil.CheckErr(o.run())
 		},
 		// PostRunE: func(cmd *cobra.Command, args []string) error {
@@ -48,7 +48,9 @@ func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOSt
 
 	cmd.Flags().StringVar(&o.ClusterPoolHost, "cph", "", "The clusterpoolhost to use")
 	cmd.Flags().StringVar(&o.outputFile, "output-file", "", "The generated resources will be copied in the specified file")
-	cmd.Flags().BoolVar(&o.SkipSchedule, "skip-schedule", false, "Set the hibernation schedule to skip")
+	cmd.Flags().BoolVar(&o.SkipSchedule, "skip-schedule", false, "Set the hibernation schedule to skip (deprecated)")
+	cmd.Flags().BoolVar(&o.HibernateScheduleOn, "hibernate-schedule-on", false, "Set the hibernation schedule to on")
+	cmd.Flags().BoolVar(&o.HibernateScheduleOff, "hibernate-schedule-off", false, "Set the hibernation schedule to off")
 
 	return cmd
 }