+++ pkg/cmd/set/clusterclaim/exec.go
@@ -13,7 +13,7 @@ var scheduleSkip string
 
 func (o *Options) complete(cmd *cobra.Command, args []string) (err error) {
 	if len(args) < 1 {
-		return fmt.Errorf("clusterclaim names are missing")
+		return fmt.Errorf("cluster names are missing")
 	}
 	o.ClusterClaims = args[0]
 	if cmd.Flags().Lookup("hibernate-schedule-on").Changed {
@@ -22,14 +22,11 @@ func (o *Options) complete(cmd *cobra.Command, args []string) (err error) {
 	if cmd.Flags().Lookup("hibernate-schedule-off").Changed {
 		scheduleSkip = "skip"
 	}
+
 	return nil
 }
 
 func (o *Options) validate(cmd *cobra.Command) error {
-	if cmd.Flags().Lookup("skip-schedule").Changed {
-		fmt.Printf("Warninfg: skip-schedule is deprecated, please use hibernate-schedule-on")
-		scheduleSkip = "skip"
-	}
 	if cmd.Flags().Lookup("hibernate-schedule-on").Changed &&
 		cmd.Flags().Lookup("hibernate-schedule-off").Changed {
 		return fmt.Errorf("flags hibernate-schedule-on and hibernate-schedule-off are mutually exclusif")
@@ -38,10 +35,11 @@ func (o *Options) validate(cmd *cobra.Command) error {
 }
 
 func (o *Options) run() (err error) {
+
 	cph, err := clusterpoolhost.GetClusterPoolHostOrCurrent(o.ClusterPoolHost)
 	if err != nil {
 		return err
 	}
 
-	return cph.HibernateClusterClaims(o.ClusterClaims, scheduleSkip, o.CMFlags.DryRun)
+	return cph.SetHibernateScheduleClusterClaims(o.ClusterClaims, scheduleSkip, o.CMFlags.DryRun)
 }